refactor: 웹 서버와 WAS 책임 분리
