+++ tomcat/src/main/java/hoony/was/RequestHandler.java
@@ -1,4 +1,7 @@
-package org.apache.coyote.http11;
+package hoony.was;
+
+import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.HttpResponse;
 
 public interface RequestHandler {
